tts/requirements.txt
piper-tts==1.3.0
fastapi==0.128.0
uvicorn==0.40.0